# edc_ogc/static/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Testbed-16 DAPA - Gridded Data",
    "contact" : {
      "name" : "<a href=\"https://www.interactive-instruments.de/\">interactive instruments</a> for <a href=\"https://www.ogc.org/projects/initiatives/t-16\">OGC Testbed-16</a>"
    },
    "license" : {
      "name" : "GHCN-Daily, see \"Special Note\"",
      "url" : "https://www.ncdc.noaa.gov/ghcnd-data-access"
    },
    "version" : "0.0.3"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/conformance" : {
      "get" : {
        "tags" : [ "Capabilities" ],
        "summary" : "conformance declaration",
        "description" : "The URIs of all conformance classes supported by the server. This information is provided to support 'generic' clients that want to access multiple OGC API implementations - and not 'just' a specific API. For clients accessing only a single API, this information is in general not relevant and the OpenAPI definition details the required information about the API.",
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "description" : "The available output formats for the response. If no value is provided, the standard HTTP rules apply, i.e., the accept header will be used to determine the format.",
          "required" : false,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "json", "html" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The operation was executed successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConformanceDeclaration"
                }
              },
              "text/html" : {
                "schema" : {
                  "$ref" : "#/components/schemas/htmlSchema"
                }
              }
            }
          },
          "400" : {
            "description" : "A query parameter has an invalid value.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/exception"
                }
              },
              "text/html" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "406" : {
            "description" : "Content negotiation failed. For example, the `Accept` header submitted in the request did not support any of the media types supported by the server for the requested resource."
          },
          "500" : {
            "description" : "A server error occurred.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/exception"
                }
              },
              "text/html" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/collections" : {
      "get" : {
        "tags" : [ "Data collections" ],
        "summary" : "feature collections in the dataset",
        "description" : "The dataset is organized in feature collections. This resource provides information about and access to the feature collections.\nThe response contains the list of collections. For each collection, a link to the items in the collection (path `/collections/{collectionId}/items`, link relation `items`) as well as key information about the collection.\nThis information includes:\n\n* A local identifier for the collection that is unique for the dataset;\n* A list of coordinate reference systems (CRS) in which geometries may be returned by the server. The first CRS is the default coordinate reference system (the default is always WGS 84 with axis order longitude/latitude);\n* An optional title and description for the collection;\n* An optional extent that can be used to provide an indication of the spatial and temporal extent of the collection - typically derived from the data;\n* An optional indicator about the type of the items in the collection (the default value, if the indicator is not provided, is 'feature').",
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "description" : "The available output formats for the response. If no value is provided, the standard HTTP rules apply, i.e., the accept header will be used to determine the format.",
          "required" : false,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "json", "html" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The operation was executed successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Collections"
                }
              },
              "text/html" : {
                "schema" : {
                  "$ref" : "#/components/schemas/htmlSchema"
                }
              }
            }
          },
          "400" : {
            "description" : "A query parameter has an invalid value.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/exception"
                }
              },
              "text/html" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "406" : {
            "description" : "Content negotiation failed. For example, the `Accept` header submitted in the request did not support any of the media types supported by the server for the requested resource."
          },
          "500" : {
            "description" : "A server error occurred.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/exception"
                }
              },
              "text/html" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/collections/{collection}/dapa/fields/" : {
      "get" : {
        "tags" : [ "DAPA", "GRID" ],
        "summary" : "Output fields/variables/properties/bands to be included in the request/processing/aggregation.\n\nThe `fields` parameter for the DAPA request can consists values either from the selected collection (e.g., all band\nnames from Sentinel-2) or declared dynamically (e.g., bands algebra NDVI=(B08-B04/B08+B04)) (see the Parameter\nsection below).",
        "parameters" : [ {
          "name" : "collection",
          "in" : "path",
          "description" : "The target collection.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "pattern" : "[\\w\\-]+",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The available field descriptions.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "properties" : {
                      "id" : {
                        "type" : "string"
                      },
                      "title" : {
                        "type" : "string"
                      },
                      "wavelength" : {
                        "type" : "number"
                      }
                    }
                  }
                },
                "example" : [ {
                  "id" : "B02",
                  "title" : "Blue",
                  "wavelength" : 492.4
                }, {
                  "id" : "B03",
                  "title" : "Green",
                  "wavelength" : 559.8
                }, {
                  "id" : "B04",
                  "title" : "Red",
                  "wavelength" : 664.6
                }, {
                  "id" : "B08",
                  "title" : "NIR",
                  "wavelength" : 832.8
                } ]
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "htmlSchema" : {
        "type" : "string",
        "example" : "<html>...</html>"
      },
      "ConformanceDeclaration" : {
        "type" : "object",
        "properties" : {
          "conformsTo" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "links" : {
            "type" : "array",
            "items" : {
              "required" : [ "href" ],
              "type" : "object",
              "properties" : {
                "title" : {
                  "type" : "string"
                },
                "rel" : {
                  "type" : "string"
                },
                "href" : {
                  "type" : "string"
                },
                "hreflang" : {
                  "type" : "string"
                },
                "templated" : {
                  "type" : "boolean"
                },
                "length" : {
                  "type" : "integer",
                  "format" : "int32"
                },
                "type" : {
                  "type" : "string"
                }
              }
            }
          },
          "description" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        }
      },
...
      "exception" : {
        "required" : [ "code" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        },
        "description" : "Information about the exception: an error code plus an optional description."
      },
...
    },
...
  }
}